chapter_001/src/main/java/ru/job4j/Calculator.java
package ru.job4j;

public class Calculator {
    private double result;

    public void add(double first, double second) {
        this.result = first + second;
    }

    public void divide(double first, double second) {
        this.result = first / second;
    }

    public void multiply(double first, double second) {
        this.result = first * second;
    }

    public void substract(double first, double second) {
        this.result = first - second;
    }

    public double getResult() {
        return this.result;
    }
}
